Fix KSkillData wire width to legacy short

+++ KncWX2Server.Net/src/KncWX2Server.Core/Common/KSkillData.cs
@@ -1,6 +1,6 @@
 namespace KncWX2Server.Core.Common;
 
-/// <summary>Legacy KSkillData: 32-bit skill id and 8-bit skill level.</summary>
+/// <summary>Legacy KSkillData: 16-bit skill id and 8-bit skill level.</summary>
 public readonly record struct KSkillData(
-    int SkillId,
+    short SkillId,
     byte SkillLevel);
